Refactor the base constants to single file
So other plugins' constant file can directly import.
