+++ cypress/utils/base_constants.js
@@ -1,0 +1,7 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export const BASE_PATH = Cypress.config('baseUrl');
+export const BACKEND_BASE = Cypress.env('openSearchUrl');
