+++ cypress/utils/constants.js
@@ -3,8 +3,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export const BASE_PATH = Cypress.config('baseUrl');
-export const BACKEND_BASE = Cypress.env('openSearchUrl');
-
+export * from './base_constants';
 export * from './plugins/anomaly-detection-dashboards-plugin/constants';
 export * from './plugins/index-management-dashboards-plugin/constants';
